[device-minion-runtime] Cleanup of event handling in master and worker minion Consolidate more definitions of memory layout into layout.h and layout.ld  SW-272 Support S-mdoe traps from U or S mode SW-317 Cleanup of machine/master/worker trap handlers to keep them as consistent as possible Initial message broadcast support - messages now include a counter to allow workers to determine if they've already handled a broadcast message Cleaned up broadcast syscall parameter encoding, lots of room for interpretation, may need to change again Fix SP init so SP points to scratch region (trap handlers will use mscratch directly, compiler code will decrement first) Packed message buffers more tightly Disable IPI_REDIRECT_PC from user mode code at startup Clean up ESR_XXX() macros to return volatile pointers Now print git version on startup Fix sscratch handling in return_from_kernel to work with recent S-mode trap handler changes Harcode some dummy kernel parameters on worker side for now Fix IPI_TRIGGER_CLEAR write so all 64 bits, not just lower 32, can be written Kernel launch loops using IPI broadcasts with software coherency working Worker minion now reserved first two stack entries for saved supervisor SP and trap scratch region - scratch still always points to scratch region Minor cleanup to MasterMinion prints, state machines are hokey but work and show the skeleton Fixed another 1U should be 1LL bug in message_send_worker()

+++ device-minion-runtime/src/WorkerMinion/src/test_kernels.c
@@ -17,7 +17,7 @@ void test_prefetch_kernel(void)
 #ifdef NO_CREDITS
     asm volatile ("wfi");
 #else
-    const unsigned int this_minion_id_bitmask = 1U << get_minion_id();
+    const uint64_t this_minion_id_bitmask = 1ULL << get_minion_id();
     //register unsigned int cycles = 0;
 
 #ifndef NO_CREDITS
